Rename pipeline.py to passes.py
