+++ xdsl/frontend/passes/desymref.py
@@ -3,7 +3,7 @@ from typing import Callable
 from xdsl.frontend import symref
 from xdsl.frontend.exception import FrontendProgramException
 from xdsl.ir import Block, MLContext, Operation, Region
-from xdsl.pipeline import ModulePass
+from xdsl.passes import ModulePass
 from xdsl.rewriter import Rewriter
 from xdsl.dialects import builtin
